+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/ChatMessageContent.java
@@ -4,7 +4,6 @@
 
 package com.azure.android.communication.chat.models;
 
-import com.azure.android.communication.common.CommunicationUserIdentifier;
 import com.azure.android.core.annotation.Fluent;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -35,11 +34,13 @@ public final class ChatMessageContent {
     private List<ChatParticipant> participants;
 
     /*
-     * Chat message content for messages of types participantAdded or
-     * participantRemoved.
+     * Identifies a participant in Azure Communication services. A participant
+     * is, for example, a phone number or an Azure communication user. This
+     * model must be interpreted as a union: Apart from rawId, at most one
+     * further property may be set.
      */
-    @JsonProperty(value = "initiator")
-    private CommunicationUserIdentifier initiator;
+    @JsonProperty(value = "initiatorCommunicationIdentifier")
+    private CommunicationIdentifierModel initiatorCommunicationIdentifier;
 
     /**
      * Get the message property: Chat message content for messages of types
@@ -108,24 +109,31 @@ public final class ChatMessageContent {
     }
 
     /**
-     * Get the initiator property: Chat message content for messages of types
-     * participantAdded or participantRemoved.
+     * Get the initiatorCommunicationIdentifier property: Identifies a
+     * participant in Azure Communication services. A participant is, for
+     * example, a phone number or an Azure communication user. This model must
+     * be interpreted as a union: Apart from rawId, at most one further
+     * property may be set.
      *
-     * @return the initiator value.
+     * @return the initiatorCommunicationIdentifier value.
      */
-    public CommunicationUserIdentifier getInitiator() {
-        return this.initiator;
+    public CommunicationIdentifierModel getInitiatorCommunicationIdentifier() {
+        return this.initiatorCommunicationIdentifier;
     }
 
     /**
-     * Set the initiator property: Chat message content for messages of types
-     * participantAdded or participantRemoved.
+     * Set the initiatorCommunicationIdentifier property: Identifies a
+     * participant in Azure Communication services. A participant is, for
+     * example, a phone number or an Azure communication user. This model must
+     * be interpreted as a union: Apart from rawId, at most one further
+     * property may be set.
      *
-     * @param initiator the initiator value to set.
+     * @param initiatorCommunicationIdentifier the
+     * initiatorCommunicationIdentifier value to set.
      * @return the ChatMessageContent object itself.
      */
-    public ChatMessageContent setInitiator(CommunicationUserIdentifier initiator) {
-        this.initiator = initiator;
+    public ChatMessageContent setInitiatorCommunicationIdentifier(CommunicationIdentifierModel initiatorCommunicationIdentifier) {
+        this.initiatorCommunicationIdentifier = initiatorCommunicationIdentifier;
         return this;
     }
 }