+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/ChatThread.java
@@ -4,7 +4,6 @@
 
 package com.azure.android.communication.chat.models;
 
-import com.azure.android.communication.common.CommunicationUserIdentifier;
 import com.azure.android.core.annotation.Fluent;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -35,10 +34,13 @@ public final class ChatThread {
     private OffsetDateTime createdOn;
 
     /*
-     * Id of the chat thread owner.
+     * Identifies a participant in Azure Communication services. A participant
+     * is, for example, a phone number or an Azure communication user. This
+     * model must be interpreted as a union: Apart from rawId, at most one
+     * further property may be set.
      */
-    @JsonProperty(value = "createdBy", required = true)
-    private CommunicationUserIdentifier createdBy;
+    @JsonProperty(value = "createdByCommunicationIdentifier", required = true)
+    private CommunicationIdentifierModel createdByCommunicationIdentifier;
 
     /*
      * The timestamp when the chat thread was deleted. The timestamp is in
@@ -110,22 +112,31 @@ public final class ChatThread {
     }
 
     /**
-     * Get the createdBy property: Id of the chat thread owner.
+     * Get the createdByCommunicationIdentifier property: Identifies a
+     * participant in Azure Communication services. A participant is, for
+     * example, a phone number or an Azure communication user. This model must
+     * be interpreted as a union: Apart from rawId, at most one further
+     * property may be set.
      *
-     * @return the createdBy value.
+     * @return the createdByCommunicationIdentifier value.
      */
-    public CommunicationUserIdentifier getCreatedBy() {
-        return this.createdBy;
+    public CommunicationIdentifierModel getCreatedByCommunicationIdentifier() {
+        return this.createdByCommunicationIdentifier;
     }
 
     /**
-     * Set the createdBy property: Id of the chat thread owner.
+     * Set the createdByCommunicationIdentifier property: Identifies a
+     * participant in Azure Communication services. A participant is, for
+     * example, a phone number or an Azure communication user. This model must
+     * be interpreted as a union: Apart from rawId, at most one further
+     * property may be set.
      *
-     * @param createdBy the createdBy value to set.
+     * @param createdByCommunicationIdentifier the
+     * createdByCommunicationIdentifier value to set.
      * @return the ChatThread object itself.
      */
-    public ChatThread setCreatedBy(CommunicationUserIdentifier createdBy) {
-        this.createdBy = createdBy;
+    public ChatThread setCreatedByCommunicationIdentifier(CommunicationIdentifierModel createdByCommunicationIdentifier) {
+        this.createdByCommunicationIdentifier = createdByCommunicationIdentifier;
         return this;
     }