+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/signaling/ConnectionState.java
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.android.communication.chat.signaling;
+
+public enum ConnectionState {
+    Unknown(0),
+    Connected(2),
+    Disconnected(3),
+    Switching(9);
+
+    private final int value;
+
+    ConnectionState(final int newValue) {
+        value = newValue;
+    }
+
+    public int getValue() { return value; }
+}