+++ sdk/communication/azure-communication-common/src/main/java/com/azure/android/communication/common/CommunicationTokenCredential.java
@@ -3,8 +3,6 @@
 
 package com.azure.android.communication.common;
 
-import com.azure.android.core.credential.AccessToken;
-
 import java.util.concurrent.Callable;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.Future;
@@ -15,7 +13,7 @@ import java.util.concurrent.TimeUnit;
  * <p>
  * This class is used to cache/refresh the access token required by Azure Communication Services.
  */
-public class CommunicationTokenCredential {
+public final class CommunicationTokenCredential {
     private UserCredential userCredential;
 
     /**
@@ -46,27 +44,28 @@ public class CommunicationTokenCredential {
     public CommunicationTokenCredential(CommunicationTokenRefreshOptions tokenRefreshOptions) {
         this.userCredential = new AutoRefreshUserCredential(
             tokenRefreshOptions.getTokenRefresher(),
-            tokenRefreshOptions.getRefreshProactively(),
+            tokenRefreshOptions.isRefreshProactively(),
             tokenRefreshOptions.getToken());
     }
 
 
     /**
-     * Get Azure core access token from credential
+     * Get communication access token from credential
      * <p>
-     * This method returns an asynchronous {@link java.util.concurrent.Future} with the AccessToken.
+     * This method returns an asynchronous {@link java.util.concurrent.Future} with the CommunicationAccessToken.
      * When the {@link CommunicationTokenCredential} is constructed with a <code>tokenRefresher</code> {@link java.util.concurrent.Callable},
-     * the AccessToken will automatically be updated as part of the {@link java.util.concurrent.Future} if the cached token exceeds the expiry threshold.
+     * the CommunicationAccessToken will automatically be updated as part of the {@link java.util.concurrent.Future} if the cached token exceeds the expiry threshold.
      * <p>
      * If this method is called after {@link #dispose()} has been invoked, a cancelled {@link java.util.concurrent.Future} will be returned.
      *
-     * @return Asynchronous {@link java.util.concurrent.Future} with the AccessToken
+     * @return Asynchronous {@link java.util.concurrent.Future} with the CommunicationAccessToken
      */
-    public Future<AccessToken> getToken() {
+    public Future<CommunicationAccessToken> getToken() {
         if (this.userCredential.isDisposed()) {
             return new CancelledTokenFuture();
         }
-        return this.userCredential.getToken();
+
+        return  this.userCredential.getToken();
     }
 
     /**
@@ -76,7 +75,7 @@ public class CommunicationTokenCredential {
         this.userCredential.dispose();
     }
 
-    private final class CancelledTokenFuture implements Future<AccessToken> {
+    private final class CancelledTokenFuture implements Future<CommunicationAccessToken> {
         @Override
         public boolean cancel(boolean mayInterruptIfRunning) {
             return false;
@@ -93,12 +92,12 @@ public class CommunicationTokenCredential {
         }
 
         @Override
-        public AccessToken get() {
+        public CommunicationAccessToken get() {
             throw new CancellationException();
         }
 
         @Override
-        public AccessToken get(long timeout, TimeUnit unit) {
+        public CommunicationAccessToken get(long timeout, TimeUnit unit) {
             throw new CancellationException();
         }
     }