+++ sdk/communication/azure-communication-common/src/main/java/com/azure/android/communication/common/CommunicationUserIdentifier.java
@@ -2,13 +2,10 @@
 // Licensed under the MIT License.
 package com.azure.android.communication.common;
 
-
-import com.azure.android.core.util.CoreUtil;
-
 /**
  * Communication identifier for Communication Services Users
  */
-public class CommunicationUserIdentifier extends CommunicationIdentifier {
+public final class CommunicationUserIdentifier extends CommunicationIdentifier {
 
     private final String id;
 
@@ -19,7 +16,7 @@ public class CommunicationUserIdentifier extends CommunicationIdentifier {
      * @throws IllegalArgumentException thrown if id parameter fail the validation.
      */
     public CommunicationUserIdentifier(String id) {
-        if (CoreUtil.isNullOrEmpty(id)) {
+        if (id == null || id.trim().isEmpty()) {
             throw new IllegalArgumentException("The initialization parameter [id] cannot be null or empty.");
         }
         this.id = id;